fix issue with long time connect to offline device

+++ backup_core/backup_thread/backup_thread.class.php
@@ -4,9 +4,9 @@
 	include_once ( __DIR__ . '/../inc/misc.php');
 
 	require_once ( __DIR__ . '/../cisco/cisco.class.php');
-
+	require_once ( __DIR__ . '/../ping/ping.class.php');	
+	
 	class backup_thread {
-
 
 		private $backup_job_file = "";
 		private $backup_job = "";
@@ -50,59 +50,61 @@
 				$device = explode('|', $device);
 				$this->backup_id = 0;
 				try {
-					$this->cisco_connection = new cisco($device[2], $device[3], $device[4], $device[1], $device[5]);
-					$config_url = strtok($device[6],':').":";
-					$config_name = strtok(':');
-
-					// Check device hostname, update if changed                                        
-					if ($device[8] !== $this->cisco_connection->cisco_hostname) {
-						$sth =$this->pdo_connection->prepare(_PDO_UPDATE_DEVICE_HOSTNAME_);
-						$sth->execute(array(':device_hostname' => $this->cisco_connection->cisco_hostname, ':device_id' => $device[0]));
-						$this->write_log($device,'NOTIFY_HOSTNAME_CHANGED '.$device[8].' => '.$this->cisco_connection->cisco_hostname);						
-					}
-					
-					// cisco->dir()
-					if (is_array($dir = $this->cisco_connection->dir($config_url))) {
-						foreach ($dir as $item) {
-							if ($item['name'] == $config_name) {
-								// get device backup
-								$sth = $this->pdo_connection->prepare(_PDO_GET_DEVICE_BACKUP_);
-								$sth->execute(array(':device_id' => $device[0],':config_datetime' => $item['datetime']));
-								$backup = $sth->fetchAll(PDO::FETCH_ASSOC);                                                            
-								// if not found 
-								if (count($backup) == 0) {
-									$dest=_BACKUP_JOB_PATH_.$this->backup_job_file.'.text';
-									// cisco->copy()
-									if ($this->cisco_connection->copy($device[6],$device[7].$this->backup_job_file.'.text')) {
-										// wait for readable 
-										while (!is_readable($dest)) { }
-										$sth = $this->pdo_connection->prepare(_PDO_INSERT_CONFIG_BACKUP_);
-										$sth->execute(array(':devices_device_id' => $device[0],
+					$p = new ping($device[1],1);
+					if ($p->pingResult['returnVar'] == 0) {
+						$this->cisco_connection = new cisco($device[2], $device[3], $device[4], $device[1], $device[5]);
+						$config_url = strtok($device[6],':').":";
+						$config_name = strtok(':');
+						// Check device hostname, update if changed                                        
+						if ($device[8] !== $this->cisco_connection->cisco_hostname) {
+							$sth =$this->pdo_connection->prepare(_PDO_UPDATE_DEVICE_HOSTNAME_);
+							$sth->execute(array(':device_hostname' => $this->cisco_connection->cisco_hostname, ':device_id' => $device[0]));
+							$this->write_log($device,'NOTIFY_HOSTNAME_CHANGED '.$device[8].' => '.$this->cisco_connection->cisco_hostname);						
+						}
+						// cisco->dir()
+						if (is_array($dir = $this->cisco_connection->dir($config_url))) {
+							foreach ($dir as $item) {
+								if ($item['name'] == $config_name) {
+									// get device backup
+									$sth = $this->pdo_connection->prepare(_PDO_GET_DEVICE_BACKUP_);
+									$sth->execute(array(':device_id' => $device[0],':config_datetime' => $item['datetime']));
+									$backup = $sth->fetchAll(PDO::FETCH_ASSOC);                                                            
+									// if not found 
+									if (count($backup) == 0) {
+										$dest=_BACKUP_JOB_PATH_.$this->backup_job_file.'.text';
+										// cisco->copy()
+										if ($this->cisco_connection->copy($device[6],$device[7].$this->backup_job_file.'.text')) {
+											// 	wait for readable 
+											while (!is_readable($dest)) { }
+											$sth = $this->pdo_connection->prepare(_PDO_INSERT_CONFIG_BACKUP_);
+											$sth->execute(array(':devices_device_id' => $device[0],
 															':config_datetime' => $item['datetime'],
 															':storage_datetime' => date(_LOG_DATE_TIME_FORMAT_),
 															':jobs_job_id' => $this->backup_job_id,												
 															':storage' => file_get_contents($dest)));
-										$this->backup_id = $this->pdo_connection->lastInsertId();
-										$this->write_log($device, 'SUCCESS');
-										unlink($dest);
+											$this->backup_id = $this->pdo_connection->lastInsertId();
+											$this->write_log($device, 'SUCCESS');
+											unlink($dest);
+										} else {
+											$this->write_log($device, 'ERROR_CISCO_COPY');
+										}
 									} else {
-										$this->write_log($device, 'ERROR_CISCO_COPY');
+										$this->backup_id = $backup[0]['backup_id'];
+										$this->write_log($device,'SUCCESS_BACKUP_ALLREADY_EXISTS');
 									}
-								} else {
-									$this->backup_id = $backup[0]['backup_id'];
-									$this->write_log($device,'SUCCESS_BACKUP_ALLREADY_EXISTS');
+									break;
 								}
-								break;
 							}
+						} else {
+								$this->write_log($device, 'ERROR_CISCO_DIR');
 						}
+						$this->cisco_connection = null;
 					} else {
-						$this->write_log($device, 'ERROR_CISCO_DIR');
+						$this->write_log($device, 'ERROR_PING');						
 					}
-					$this->cisco_connection = null;
-
+					$p = null;
 				} catch (Exception $e) {
 					$this->write_log($device , 'ERROR_CISCO_EXCEPTION: '.$e->getMessage());
-					return;
 				}
 
 			}
